requirements/requirements.txt
Flask>=2.3,<3.0
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
python-dotenv>=1.0
Flask-Login>=0.6
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.2
